+++ app/Http/Controllers/BuildingController.php
@@ -8,51 +8,69 @@ use App\building;
 class BuildingController extends Controller
 {
     public function view(){
-        $list = building::all()->sortBy('id');  
-        return view('',compact('list'));
+        if(Auth::check()){
+            $list = building::all()->sortBy('id');  
+            return view('',compact('list'));
+        }
+        else return redirect('/');
     }
     public function create(){
-        return view('');
+        if(Auth::check()){
+           return view(''); 
+        }
+        else return redirect('/');
     }
     public function store(Request $request){
-        $input = new building();
-        $data = $this->validate($request, [
-            'nama_gedung'=>'required',
-            'jumlah_lantai'=>'required',
-            ]);
-        $input->nama_gedung=$data['nama_gedung'];
-        $input->jumlah_lantai=$data['jumlah_lantai'];
-        $input->save();
+        if(Auth::check()){
+            $input = new building();
+            $data = $this->validate($request, [
+                'nama_gedung'=>'required',
+                'jumlah_lantai'=>'required',
+                ]);
+            $input->nama_gedung=$data['nama_gedung'];
+            $input->jumlah_lantai=$data['jumlah_lantai'];
+            $input->save();
 
-        return redirect('');
+            return redirect('');
+        }
+        else return redirect('/');
     }
     public function updatepick($id){
-        $buildings = building::where('id',$id)->first();
-        if ($buildings != null){
-            return view('',compact('buildings'));
+        if(Auth::check()){
+            $buildings = building::where('id',$id)->first();
+            if ($buildings != null){
+                return view('',compact('buildings'));
+            }
+            else return back();
         }
-        else return back();
+        else return redirect('/');
     }
 
     public function update(Request $request){
-        $input = building::where('id',$request['id'])->first();
-        $data = $this->validate($request, [
-            'nama_gedung'=>'required',
-            'jumlah_lantai'=>'required',
-            ]);
-        $input->nama_gedung=$data['nama_gedung'];
-        $input->jumlah_lantai=$data['jumlah_lantai'];
-        $input->save();
+        if(Auth::check()){
+          $input = building::where('id',$request['id'])->first();
+          $data = $this->validate($request, [
+              'nama_gedung'=>'required',
+              'jumlah_lantai'=>'required',
+              ]);
+          $input->nama_gedung=$data['nama_gedung'];
+          $input->jumlah_lantai=$data['jumlah_lantai'];
+          $input->save();
 
-        return redirect('');
+          return redirect('');  
+        }
+        return redirect('/');
     }
 
     public function delete(Request $request){
-        $buildings = building::where('id',$request['id'])->first();
-        if ($buildings != null){
-             $buildings->delete();
-        }       
-        return back();
+        if(Auth::check()){
+          $buildings = building::where('id',$request['id'])->first();
+          if ($buildings != null){
+               $buildings->delete();
+          }       
+          return back();  
+        }
+        else return redirect('/');
     }
